+++ qcommon/pmove.c
@@ -34,7 +34,7 @@ typedef struct
 	vec3_t		velocity;		// full float precision
 
 	vec3_t		forward, right, up;
-	double		frametime;
+	float		frametime;
 
 
 	csurface_t	*groundsurface;
