+++ win32/sys_win.c
@@ -43,8 +43,8 @@ qboolean	Minimized;
 
 static HANDLE		hinput, houtput;
 
-double	sys_msg_time;
-double	sys_frame_time;
+unsigned int	sys_msg_time;
+unsigned int	sys_frame_time;
 
 
 static HANDLE		qwclsemaphore;
@@ -314,7 +314,6 @@ char *Sys_ConsoleInput (void)
 						}
 
 						break;
-
 				}
 			}
 		}
@@ -354,6 +353,7 @@ void Sys_ConsoleOutput (char *string)
 		WriteFile(houtput, console_text, console_textlen, &dummy, NULL);
 }
 
+
 /*
 ================
 Sys_Sleep
@@ -365,6 +365,7 @@ void Sys_Sleep (unsigned msec)
 {
 	Sleep (msec);
 }
+
 
 /*
 ================
@@ -381,15 +382,14 @@ void Sys_SendKeyEvents (void)
 	{
 		if (!GetMessage (&msg, NULL, 0, 0))
 			Sys_Quit ();
-		sys_msg_time = (double)msg.time;
+		sys_msg_time = msg.time;
       	TranslateMessage (&msg);
       	DispatchMessage (&msg);
 	}
 
 	// grab frame time 
-	sys_frame_time = (double)timeGetTime();	// FIXME: should this be at start?
+	sys_frame_time = timeGetTime();	// FIXME: should this be at start?
 }
-
 
 
 /*
@@ -645,7 +645,7 @@ HINSTANCE	global_hInstance;
 int WINAPI WinMain (HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nCmdShow)
 {
 	MSG			msg;
-	double				time, oldtime, newtime;
+	int				time, oldtime, newtime;
 	char			*cddir;
 
     /* previous instances do not exist in Win32 */
@@ -691,7 +691,7 @@ int WINAPI WinMain (HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLin
 		{
 			if (!GetMessage (&msg, NULL, 0, 0))
 				Com_Quit ();
-			sys_msg_time = (double)msg.time;
+			sys_msg_time = msg.time;
 			TranslateMessage (&msg);
 			DispatchMessage (&msg);
 		}
